# Introduce ENFORCE_FAST to speed up tests in debug builds