--- a/core/Symbols.h
+++ b/core/Symbols.h
@@ -190,52 +190,52 @@ public:
     }
 
     inline bool isOverloaded() const {
-        ENFORCE(isMethod());
+        ENFORCE_FAST(isMethod());
         return (flags & Symbol::Flags::METHOD_OVERLOADED) != 0;
     }
 
     inline bool isAbstract() const {
-        ENFORCE(isMethod());
+        ENFORCE_FAST(isMethod());
         return (flags & Symbol::Flags::METHOD_ABSTRACT) != 0;
     }
 
     inline bool isIncompatibleOverride() const {
-        ENFORCE(isMethod());
+        ENFORCE_FAST(isMethod());
         return (flags & Symbol::Flags::METHOD_INCOMPATIBLE_OVERRIDE) != 0;
     }
 
     inline bool isGenericMethod() const {
-        ENFORCE(isMethod());
+        ENFORCE_FAST(isMethod());
         return (flags & Symbol::Flags::METHOD_GENERIC) != 0;
     }
 
     inline bool isOverridable() const {
-        ENFORCE(isMethod());
+        ENFORCE_FAST(isMethod());
         return (flags & Symbol::Flags::METHOD_OVERRIDABLE) != 0;
     }
 
     inline bool isOverride() const {
-        ENFORCE(isMethod());
+        ENFORCE_FAST(isMethod());
         return (flags & Symbol::Flags::METHOD_OVERRIDE) != 0;
     }
 
     inline bool isCovariant() const {
-        ENFORCE(isTypeArgument() || isTypeMember());
+        ENFORCE_FAST(isTypeArgument() || isTypeMember());
         return (flags & Symbol::Flags::TYPE_COVARIANT) != 0;
     }
 
     inline bool isInvariant() const {
-        ENFORCE(isTypeArgument() || isTypeMember());
+        ENFORCE_FAST(isTypeArgument() || isTypeMember());
         return (flags & Symbol::Flags::TYPE_INVARIANT) != 0;
     }
 
     inline bool isContravariant() const {
-        ENFORCE(isTypeArgument() || isTypeMember());
+        ENFORCE_FAST(isTypeArgument() || isTypeMember());
         return (flags & Symbol::Flags::TYPE_CONTRAVARIANT) != 0;
     }
 
     inline bool isFixed() const {
-        ENFORCE(isTypeArgument() || isTypeMember());
+        ENFORCE_FAST(isTypeArgument() || isTypeMember());
         return (flags & Symbol::Flags::TYPE_FIXED) != 0;
     }
 
@@ -250,22 +250,22 @@ public:
     }
 
     inline bool isPublic() const {
-        ENFORCE(isMethod());
+        ENFORCE_FAST(isMethod());
         return !isProtected() && !isPrivate();
     }
 
     inline bool isProtected() const {
-        ENFORCE(isMethod());
+        ENFORCE_FAST(isMethod());
         return (flags & Symbol::Flags::METHOD_PROTECTED) != 0;
     }
 
     inline bool isPrivate() const {
-        ENFORCE(isMethod());
+        ENFORCE_FAST(isMethod());
         return (flags & Symbol::Flags::METHOD_PRIVATE) != 0;
     }
 
     inline bool isClassOrModuleModule() const {
-        ENFORCE(isClassOrModule());
+        ENFORCE_FAST(isClassOrModule());
         if (flags & Symbol::Flags::CLASS_OR_MODULE_MODULE)
             return true;
         if (flags & Symbol::Flags::CLASS_OR_MODULE_CLASS)
@@ -274,7 +274,7 @@ public:
     }
 
     inline bool isClassModuleSet() const {
-        ENFORCE(isClassOrModule());
+        ENFORCE_FAST(isClassOrModule());
         return flags & (Symbol::Flags::CLASS_OR_MODULE_MODULE | Symbol::Flags::CLASS_OR_MODULE_CLASS);
     }
 
@@ -288,22 +288,22 @@ public:
     }
 
     inline bool isClassOrModuleInterface() const {
-        ENFORCE(isClassOrModule());
+        ENFORCE_FAST(isClassOrModule());
         return (flags & Symbol::Flags::CLASS_OR_MODULE_INTERFACE) != 0;
     }
 
     inline bool isClassOrModuleLinearizationComputed() const {
-        ENFORCE(isClassOrModule());
+        ENFORCE_FAST(isClassOrModule());
         return (flags & Symbol::Flags::CLASS_OR_MODULE_LINEARIZATION_COMPUTED) != 0;
     }
 
     inline bool isClassOrModuleFinal() const {
-        ENFORCE(isClassOrModule());
+        ENFORCE_FAST(isClassOrModule());
         return (flags & Symbol::Flags::CLASS_OR_MODULE_FINAL) != 0;
     }
 
     inline bool isClassOrModuleSealed() const {
-        ENFORCE(isClassOrModule());
+        ENFORCE_FAST(isClassOrModule());
         return (flags & Symbol::Flags::CLASS_OR_MODULE_SEALED) != 0;
     }