<!-- (optional) Explain your change, focusing on the details of the solution. This is a great place to call out user-visible changes. -->

Introduce ENFORCE_FAST to speed up tests in debug builds.

ENFORCE_FAST is equal to ENFORCE, except it doesn't create timers.


### Motivation
<!-- Why make this change? Describe the problem, not the solution. This can also be a link to an issue. -->

Timers were making debug test runs >50% slower locally. With these changes, test/testdata/lsp/definitions_and_usages takes less than half as much time as it used to (7s-8s to ~3.3s on my laptop).

* 7.37s - 8.12s Before optimizations
* 5.58s - 6.0s after SymbolRef::data/dataAllowingNone change
* 5.3 - 5.6 after resolver::computeClassLinearization change
* 4.0s - 4.4s after SymbolDataDebugCheck::check change
* 4.0s after Symbol::mixins(), superClass()
* 3.8s after Symbol::sanityCheck
* 3.5s after GS::enterSymbol >2.7%
* 3.4s after Symbol::arguments and a few others ~0.7%
* ~3.25s after Name::sanityCheck / gs::sanityCheck

Timers still show up prominently in profiles, and I know there's probably another ~500ms of savings to trim, but this is good enough for now and will speed up my workflow.

### Test plan
<!-- If you did not write tests for this change, replace the message below explaining why not. Why we should be confident this change is correct? If you changed the website, please include a screenshot of the proposed changes. -->

This change does not impact correctness.