@jvilk-stripe 2020-08-04
I am open to renaming the new ENFORCE directive, too (ENFORCE_NO_TIMER?).